refactor(package): インストール状態の判定パイプラインを main プロセスへ移設
- services/packages.ts に getInstalledFiles / getPackagesExtra /
  getPackagesWithStatus を追加。tRPC の packages.getPackagesExtra /
  packages.getPackagesWithStatus query として公開
- getPackagesWithStatus は旧 setPackagesList 前半(取得 → 状態付与 →
  整合性による apm.json 補正 → 依存解決)と同一の挙動。
  fixIntegrity=false は旧 installScript の redirect 解決経路と同一
- 全呼び出し元が getPackages の新鮮な結果を渡していたため、
  main 側で一覧取得から行う instPath のみの引数に変更
  (setPackagesList / installScript / displayNicommonsIdList / batchInstall)
- renderer 側 packageUtil の getInstalledFiles / getPackagesStatus は
  main へ移設したため削除
